Add separate `context_server_settings` crate
