+++ crates/context_server/src/context_server.rs
@@ -7,11 +7,10 @@ mod registry;
 pub mod types;
 
 use command_palette_hooks::CommandPaletteFilter;
+pub use context_server_settings::{ContextServerSettings, ServerCommand, ServerConfig};
 use gpui::{actions, AppContext};
-use settings::Settings;
 
 pub use crate::context_server_tool::ContextServerTool;
-use crate::manager::ContextServerSettings;
 pub use crate::registry::ContextServerFactoryRegistry;
 
 actions!(context_servers, [Restart]);
@@ -20,7 +19,7 @@ actions!(context_servers, [Restart]);
 pub const CONTEXT_SERVERS_NAMESPACE: &'static str = "context_servers";
 
 pub fn init(cx: &mut AppContext) {
-    ContextServerSettings::register(cx);
+    context_server_settings::init(cx);
     ContextServerFactoryRegistry::default_global(cx);
     extension_context_server::init(cx);
