+++ crates/context_server_settings/LICENSE-GPL
@@ -1,0 +1,1 @@
+../../LICENSE-GPL
