+++ crates/extension_host/Cargo.toml
@@ -22,7 +22,7 @@ async-tar.workspace = true
 async-trait.workspace = true
 client.workspace = true
 collections.workspace = true
-context_servers.workspace = true
+context_server_settings.workspace = true
 extension.workspace = true
 fs.workspace = true
 futures.workspace = true
